requirements.txt
# --- SERVEUR & API ---
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
jinja2==3.1.3

# --- TRAITEMENT PDF (Optimisé) ---
pypdf==4.0.1
pdf2docx==0.5.6
pdf2image==1.17.0
pdfplumber==0.10.3
pikepdf==8.13.0
reportlab==4.0.9

# --- TRAITEMENT IMAGES & SCAN ---
Pillow==10.2.0
pytesseract==0.3.10
opencv-python-headless==4.9.0.80

# --- OFFICE & DATA ---
python-docx==1.1.0
python-pptx==0.6.21
XlsxWriter==3.1.9
openpyxl==3.1.2
pandas==2.2.0
tabula-py==2.9.0
ocrmypdf==16.1.2

# --- GÉNÉRATION DE DOCUMENTS & HTML ---
requests==2.31.0
weasyprint==61.2

# --- UTILITAIRES & SÉCURITÉ ---
typing-extensions==4.9.0
fonttools==4.47.2
cryptography==42.0.2

# --- MÉTADONNÉES ---
exifread==3.0.0
# Authentification
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Base de données
sqlalchemy==2.0.25
alembic==1.13.0

# Email
sendgrid==6.10.0
aiosmtplib==2.0.2

# Cloud
google-auth==2.25.0
dropbox==11.36.2

# Paiements
stripe==7.5.0
# --- SUPPRIMÉS (Économie de mémoire) ---
# - camelot-py (trop lourd, 300MB)
# - pdfminer.six (déjà dans pdfplumber)
# - openai (à activer si besoin)
# - redis (non utilisé)
# - sendgrid (non utilisé)
# - google-* (non utilisé)
# - pdf-diff (non utilisé)